chore: combine error handling into existing extension
