+++ src/xcode/ENA/ENA/Source/Scenes/ExposureSubmission/ExposureSubmissionTestResultViewController.swift
@@ -134,7 +134,7 @@ class ExposureSubmissionTestResultViewController: DynamicTableViewController, EN
 				switch result {
 				case let .failure(error):
 
-					let alert = ErrorAlertUtil.setupErrorAlert(message: error.localizedDescription)
+					let alert = self.setupErrorAlert(message: error.localizedDescription)
 					
 					self.present(alert, animated: true, completion: {
 						self.navigationFooterItem?.isPrimaryButtonEnabled = true
@@ -159,7 +159,7 @@ class ExposureSubmissionTestResultViewController: DynamicTableViewController, EN
 		if let state = exposureSubmissionService?.preconditions() {
 			if !state.isGood {
 
-				let alert = ErrorAlertUtil.setupErrorAlert(
+				let alert = self.setupErrorAlert(
 					message: ExposureSubmissionError.enNotEnabled.localizedDescription
 				)
 				self.present(alert, animated: true, completion: nil)
